requirements.txt
Flask
nltk
scikit-learn
PyPDF2
python-docx
